+++ kwimage/im_cv2.py
@@ -992,8 +992,8 @@ def convert_colorspace(img, src_space, dst_space, copy=False,
             lab = convert_colorspace(img01, 'rgb', 'lab')
             np.minimum(lab[0, 0], minvals, out=minvals)
             np.maximum(lab[0, 0], maxvals, out=maxvals)
-        print('minvals = {}'.format(ub.repr2(minvals, nl=0)))
-        print('maxvals = {}'.format(ub.repr2(maxvals, nl=0)))
+        print('minvals = {}'.format(ub.urepr(minvals, nl=0)))
+        print('maxvals = {}'.format(ub.urepr(maxvals, nl=0)))
     """
     src_space = src_space.upper()
     dst_space = dst_space.upper()
@@ -1925,7 +1925,7 @@ def _prepare_downscale(image, sx, sy):
             loss = ((got - want) ** 2).sum()
             return loss
         result = scipy.optimize.minimize(sigma_error, x0=1.0, method='Nelder-Mead')
-        print('result = {}'.format(ub.repr2(result, nl=1)))
+        print('result = {}'.format(ub.urepr(result, nl=1)))
         # This gives a number like 1.06992187 which is not exactly what
         # we use.
         #
@@ -1994,7 +1994,7 @@ def _pyrDownK(a, k=1):
 """
 items = {k.split('_')[1].lower(): 'cv2.' + k for k in dir(cv2) if k.startswith('MORPH_')}
 items = ub.sorted_vals(items, key=lambda x: eval(x, {'cv2': cv2}))
-print('_CV2_MORPH_MODES = {}'.format(ub.repr2(items, nl=1, sv=1, align=':')))
+print('_CV2_MORPH_MODES = {}'.format(ub.urepr(items, nl=1, sv=1, align=':')))
 """
 _CV2_STRUCT_ELEMENTS = {
     'rect'    : cv2.MORPH_RECT,
@@ -2103,7 +2103,7 @@ def morphology(data, mode, kernel=5, element='rect', iterations=1,
         >>> grid += [{'mode': 'dilate', 'kernel': 'random', 'element': 'custom'}]
         >>> results = {}
         >>> for params in grid:
-        ...     key = ub.repr2(params, compact=1, si=0, nl=1)
+        ...     key = ub.urepr(params, compact=1, si=0, nl=1)
         ...     if params['kernel'] == 'random':
         ...         params['kernel'] = np.random.rand(5, 5)
         ...     results[key] = morphology(image, **params)